requirements.txt
google-cloud-bigquery==3.*
discord==2.*